+++ frontend/index.html
@@ -11,3 +11,8 @@
     <script type="module" src="./src/main.js"></script>
   </body>
 </html>
+<style>
+  #nprogress .spinner {
+    display: none; /* 进度条加载的右边的圆圈让不显示 */
+  }
+</style>
