+++ frontend/src/views/PublicChoice.vue
@@ -50,7 +50,7 @@
 </template>
 
 <script setup>
-import { ref, onMounted } from 'vue'
+import { ref, onMounted, inject } from 'vue'
 import Paginator from "@/components/Paginator.vue";
 import Drawer from "@/components/Drawer.vue";
 import SearchBar from "@/components/SearchBar.vue";
@@ -79,6 +79,8 @@ const updateCourseData = async (page) => {
 }
 
 onMounted(async () => {
+    const prepare = inject('prepare')
+    await prepare()
     await updateCourseData(1)
 })
